==> scripts/PICI_microbiome/scripts/01_clinical_table.Rmd <==
---
title: "The clinical table"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = F)
```

```{r}
library(tidyverse)
library(ggpubr)
```

AMADEUS has 41 patients. 

```{r}
ptb <- read_csv('../data/input/PICI0025_clinical_2021-10-07/AMADEUS_ClinicalData_20211007.csv')

patient <- ptb %>% 
  rename_all(~ gsub(" ", "_", .)) %>% 
  select(pid = Subject, Manual_Tumor_Buckets, Notes_for_Informatics, Arm:DCR_Responder)

patient %>% write_csv('../data/01_patient_table_long.csv')


patient %>% 
  select(pid:Study_Day_of_Crossover, Best_Overall_Response:DCR_Responder) %>% 
  write_csv('../data/01_patient_table_short.csv')

ptbs <- patient %>% 
  select(pid:Study_Day_of_Crossover, Best_Overall_Response:DCR_Responder)
```


```{r}
stb <- readxl::read_excel('/Volumes/vandenBrinkLab/PICI/PICI_microbiome_analysis_project/PICI0025_PICI0033_microbiome_samples_shipped.xlsx') %>% 
  rename_all(~ gsub(" ", "_", .)) %>% 
  select(sampleid = Sample_ID,
         Participant_ID,
         Visit, 
         Study_Day) %>% 
  mutate(pid = str_extract(Participant_ID, '-1.+$')) %>% 
  mutate(pid = str_replace(pid, '^-','')) %>% 
  select(-Participant_ID) %>% 
  mutate(Visit = str_replace(Visit, ' ','_'))

stb %>% write_csv('../data/01_sample_table.csv')
# the metrics from the pipeline
metric <- read_tsv('../data/input/readcounts.tsv')
```

```{r}
# the meta table that has both ptb and stb info
meta_amadeus <- stb %>% 
  inner_join(ptbs, by = "pid")

meta_amadeus %>% 
  write_csv('../data/01_meta_amadeus.csv')


cohortA <- read_csv('../data/input/PICI0033_CohortA_clinical_2021-02-16/PorterA_ClinicalData_20210216.csv')

library(ggvenn)

two_list <- list(amadeus_pt = ptbs$pid, allsamps_pt = stb$pid, porterA_pt = cohortA$Subject)
ggvenn(two_list, show_percentage = F)
ggsave('../figs/01_samples_venn.png', width = 4, height = 4)

# do we have all the patients 
```


```{r}
# visualize the read depth of the shotgun data
# what proportions of the reads left after all each preprocessing steps
metric %>% 
  select(ends_with('_frac')) %>% 
  select(-orphan_frac) %>% 
  gather('Group', 'Fraction') %>% 
  mutate(Group = str_replace(Group, '_frac','')) %>% 
  ggboxplot( x = 'Group', y = 'Fraction', add = 'jitter',add.params = list(alpha = 0.7),
             ylab = 'Fraction of reads left after each step',
             xlab = '') +
  theme_classic()

ggsave('../figs/01_fraction_reads_preprocessing.pdf', width = 4, height = 3)
```


```{r}
# visualize the surviving read depth
our <- read_tsv('../data/input/2021-11-10-knead_stats.tab')
both <- bind_rows(
  tibble(center = 'MSK', depth = our$filt_readsr),
  tibble(center = 'PICI', depth = metric$host_removed_reads)
) 

both %>% 
  ggboxplot(x = 'center', y = 'depth', add = 'jitter' , add.params = list(alpha = 0.3),
            title = 'Read counts after preprocessing',
            ylab = 'Total number of reads for each sample')  +
  geom_hline(yintercept = 3e7, linetype = 'dashed') 

ggsave('../figs/01_read_depth_both.pdf', width = 4, height = 3)

median(metric$host_removed_reads)
median(our$filt_readsr)

```
